Try to fix SetProcessTitle, once more
Found some dark magic we can use to make the Linux kernel take the
strlen of the title and only use that.  We could do a quick peek into
/proc/<pid>/cmdline to check that it worked and revert to something
different if it didn't, but I didn't bother doing that.  I've already
spent way too much energy on something this small.